FEAT: Filtros categóricos e de busca funcionais na pagina home, extenções .js dos arquivos styles trocados para .jsx
